refactor: TableGroup 도메인에서 OrderTable 의존성 제거
